+++ tests/testthat/helper-all.R
@@ -35,3 +35,13 @@ file_name <- function(object, x32 = FALSE) {
     else
         paste0("rds/", as.character(substitute(object)), ".rds")
 }
+
+# expect_equal for specific slots
+expect_equal_slots <- function(current, target, slots = c("cluster", "centroids", "cldist"), ...) {
+    for (object_slot in slots) {
+        expect_equal(slot(current, object_slot),
+                     slot(target, object_slot),
+                     ...,
+                     info = paste("slot =", object_slot))
+    }
+}