+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Version 5.2.0.9000
+## Version 5.3.0
 * Added 2 shiny apps, see `interactive_clustering` and `ssdtwclust`.
 * Added another helper for `compare_clusterings`: `repeat_clustering`.
 * Added `distmat` to `fuzzy_control` for when `centroid = "fcmdd"`.